Add restart functionality
This resets the level and fades the music back in if it stopped. Note that require a slight rework of the DelayedRunnableHandler class to avoid concurrentthis required a slight rework of the DelayedRunnableHandler internals to avoid concurrent modification exceptions when a runnable added another to the queue while running itself.